requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15.2
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==1.1.3
h11==0.13.0
idna==3.4
multidict==6.0.2
outcome==1.2.0
psycopg2==2.9.3
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
yarl==1.8.1
